update: see details
1) fix sysio and pci field offset bugs
2) acpi reset and poweroff
3) qemu reboots with dirty memory. on efi clean all pages
4) fix bit fields bugs.

+++ includes/fat.h
@@ -83,15 +83,15 @@ typedef struct {
 #define FAT32_DIRENT_TYPE_UNUSED      0xE5
 
 typedef struct {
-	uint8_t seconds : 5;
-	uint8_t minutes : 6;
-	uint8_t hours  : 5;
+	uint16_t seconds : 5;
+	uint16_t minutes : 6;
+	uint16_t hours  : 5;
 }__attribute__((packed)) fat32_dirent_time;
 
 typedef struct {
-	uint8_t day : 5;
-	uint8_t month : 4;
-	uint8_t year : 7;
+	uint16_t day : 5;
+	uint16_t month : 4;
+	uint16_t year : 7;
 }__attribute__((packed)) fat32_dirent_date;
 
 typedef struct {
